Consolidates helper functions into lib/values.js
