--- lib/hasEmptyValue.js
@@ -1,18 +1,0 @@
-/**
- * Returns true for values that contain an empty or otherwise useless value.
- *
- * @param  {Mixed} values The value to be checked.
- * @return {Boolean}      Whether the value is considered empty.
- */
-const hasEmptyValue = (value) => (
-  [
-    false,
-    'false',
-    null,
-    undefined,
-    0,
-    '0',
-  ].includes(value)
-);
-
-module.exports = hasEmptyValue;
