+++ lib/transformValue.js
@@ -1,7 +1,7 @@
 const valueParser = require('postcss-value-parser');
 const reducer = require('postcss-calc/dist/lib/reducer');
 const { parser } = require('postcss-calc/dist/parser');
-const roundToPrecision = require('./roundToPrecision');
+const { roundToPrecision } = require('./values');
 
 /**
  * Stringify the reduced AST.
